修复#8880没有考虑好同样的菜品点了多次，然后使用赠菜券导致数量计算问题。 发布Beta5.

+++ CanDao.Pos.Model/DishGiftCouponInfo.cs
@@ -13,7 +13,7 @@
         /// <summary>
         /// 使用的优惠券数量。
         /// </summary>
-        public int UsedCouponCount { get; set; }
+        public decimal UsedCouponCount { get; set; }
 
     }
 }
